+++ go/protocol/keybase1/user.go
@@ -302,18 +302,6 @@ func (o ProofSuggestion) DeepCopy() ProofSuggestion {
 	}
 }
 
-type SizedImage struct {
-	Path  string `codec:"path" json:"path"`
-	Width int    `codec:"width" json:"width"`
-}
-
-func (o SizedImage) DeepCopy() SizedImage {
-	return SizedImage{
-		Path:  o.Path,
-		Width: o.Width,
-	}
-}
-
 type NextMerkleRootRes struct {
 	Res *MerkleRootV2 `codec:"res,omitempty" json:"res,omitempty"`
 }
