+++ protocol/avdl/keybase1/identify3_ui.avdl
@@ -50,7 +50,7 @@ protocol identify3Ui {
     string key,                    // e.g., "twitter", "web", "github", "btc", "pgp"
     string value,                  // e.g., "bob", "foo.com", "1BYzrCvfbn81dfiksmD1Bdgt8pgLi1SD7Z", "8EFBE2E4DD56B35273634E8F6052B2AD31A6631C"
     string siteURL,                // https://twitter.com/bob
-    string siteIcon,               // https://keybase.io/_/icons/twitter.png
+    array<SizedImage> siteIcon,
     string proofURL,               // http://twitter.com/bob/post/1234
     SigID sigID,                   // so we can later revoke the sig....
     Identify3RowState state,       // 'checking' | 'valid' | 'error' | 'warning' | 'revoked',
