+++ protocol/json/keybase1/identify3_ui.json
@@ -99,7 +99,10 @@
         },
         {
           "name": "siteIcon",
-          "type": "string"
+          "type": {
+            "type": "array",
+            "items": "SizedImage"
+          }
         },
         {
           "name": "proofURL",
