+++ protocol/json/keybase1/user.json
@@ -301,20 +301,6 @@
     },
     {
       "type": "record",
-      "name": "SizedImage",
-      "fields": [
-        {
-          "type": "string",
-          "name": "path"
-        },
-        {
-          "type": "int",
-          "name": "width"
-        }
-      ]
-    },
-    {
-      "type": "record",
       "name": "NextMerkleRootRes",
       "fields": [
         {
